+++ src/graph/mesh/SphereMesh.cpp
@@ -8,7 +8,7 @@
 namespace artemis::graph {
 
 void SphereMesh::on_process() {
-	GeometrySphere geo(v_0, radius(), samples());
+	GeometrySphere geo(v_0, (float)radius(), samples());
 	geo.smoothen();
 	out(geo);
 }
